Add tests for successful data retrieval
Note that due to changes in how FB uses the me/music endpoint and in how
they allow test users to interact with the sandboxed instance of
Facebook, test users cannot populate their music likes.
As such, a web requestor is written to spoof responses to (and only to)
that endpoint, returning hardcoded responses for specific user IDs.

Additionally, location data is not consistent. That means that these
tests may not currently be passing, but they should pass once enough
time has passed from the posting of the tagged places.